+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Added support for `DatetimeIndex.mean` and `DatetimeIndex.std` methods.
 - Added support for `Resampler.asfreq`.
 - Added support for `resample` frequency `W`, `ME`, `YE` with `closed = "left"`.
+- Added support for `DataFrame.rolling.corr` and `Series.rolling.corr` for `pairwise = False` and int `window`.
 - Added support for string time-based `window` for `Rolling`.
 
 #### Improvements